requirements.txt
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cmdstanpy==0.9.5
convertdate==2.2.2
cycler==0.10.0
Cython==0.29.21
dash==1.16.1
dash-bootstrap-components==0.10.6
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.1
dash-table==4.10.1
dill==0.3.2
ephem==3.7.7.1
fbprophet==0.7.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
holidays==0.10.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.2
Pillow==7.2.0
plotly==4.10.0
psycopg2-binary==2.8.6
PyMeeus==0.3.7
pyparsing==2.4.7
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
setuptools-git==1.2
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.20
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.0
urllib3==1.25.10
Werkzeug==1.0.1
